backend/requirements.txt
Django==4.2.19
django-environ==0.12.0
psycopg2==2.9.10
pillow==11.1.0
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.4.0
